Refactor Solana utilities to use @solana/kit packages
- Replace custom base58 encode/decode with @solana/codecs
- Replace custom address validation with @solana/addresses
- Remove ~200 lines of custom crypto code from lib/solana.ts
- Keep only getSolanaConfig() and a thin assertValidAddress wrapper
- Add @solana/codecs and @solana/addresses to dependencies
- Fix OpenAPI spec type inference issue

The custom implementations had a bug (isOnCurve always returned false)
and duplicated functionality already provided by the official SDK.
