+++ apps/sdp-api/package.json
@@ -27,6 +27,8 @@
     "@sdp/types": "workspace:*",
     "@solana-program/system": "catalog:",
     "@solana-program/token-2022": "catalog:",
+    "@solana/addresses": "catalog:",
+    "@solana/codecs": "catalog:",
     "@solana/kit": "catalog:",
     "hono": "^4.7.0",
     "nanoid": "^5.1.0",
@@ -37,6 +39,7 @@
     "@cloudflare/vitest-pool-workers": "^0.12.0",
     "@cloudflare/workers-types": "^4.20250124.0",
     "@vitest/coverage-v8": "^3.0.0",
+    "openapi3-ts": "4.5.0",
     "tsx": "4.21.0",
     "typescript": "^5.8.0",
     "vitest": "^3.0.0",
